Move JsonNamingPolicy to Core
